directly pass {g,gw}_idxs
They were previously passed together in a common dataset, but I had a weird bug, not happening all the time, where win_dim dimension would be reduced to the smallest size, the one g_idxs. But gw_idxs is supposed to have a larger dimension.

Perhaps the problem is that g_idxs should have the same win_dim as gw_idxs, but fill all irrelevant entries with nans. This would avoid the problem. Perhaps in the future I should try this instead. For now it works